Entity Framework Implementation Skeleton Changeset highlights: New solution added (KiGGEF.sln) Solution is structured with solution folders New projects added: Kigg.Infrastructure.EF (EF Implementation Repositories & Models) Kigg.Infrastructure.EF.Test (xUnit Tests) Kigg.Infrastructure.EF.IntegrationTest (Integration Test)
All EF based Models are not implemented yet
All EF based Repositories are not implemented

ObjectContext for EF is implemented with Test.
DataLoadOptions for EF is impelemented with Test.

EF implementation is not using VS.NET EDM Designer. CSDL, SSDL and MSL are with desiner then extracted and modified manually.